[add] new system

+++ src/systems/render.c
@@ -25,7 +25,7 @@ void Scroll_Text(tnecs_input *input) {
     struct Timer     *ut_arr = TNECS_COMPONENT_ARRAY(input, Timer_ID);
     struct PixelFont *pf_arr = TNECS_COMPONENT_ARRAY(input, PixelFont_ID);
 
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Timer     *ut = (ut_arr + order);
         struct PixelFont *pf = (pf_arr + order);
 
@@ -39,7 +39,7 @@ void Animate_Sprite(tnecs_input *input) {
     struct Timer  *ut_arr     = TNECS_COMPONENT_ARRAY(input, Timer_ID);
     struct Sprite *sprite_arr = TNECS_COMPONENT_ARRAY(input, Sprite_ID);
 
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Timer       *ut     = (ut_arr     + order);
         struct Sprite      *sprite = (sprite_arr + order);
 
@@ -73,7 +73,7 @@ void Draw_Actor(tnecs_input *input) {
     SDL_assert(actor_arr   != NULL);
     SDL_assert(position_arr != NULL);
 
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Actor    *actor      = (actor_arr   + order);
         struct Position *position   = (position_arr   + order);
 
@@ -106,7 +106,7 @@ void Draw_Sprite(tnecs_input *input) {
     tnecs_world   *world        = input->world;
     size_t         archetype_id = input->entity_archetype_id;
 
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Sprite   *sprite   = (sprite_arr   + order);
 
         if (!sprite->visible)
@@ -139,7 +139,7 @@ void Draw_Map_Boss_Icon(tnecs_input *input) {
     struct Sprite   *sprite_arr     = TNECS_COMPONENT_ARRAY(input, Sprite_ID);
     struct Position *position_arr   = TNECS_COMPONENT_ARRAY(input, Position_ID);
 
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Boss     *boss       = (boss_arr     + order);
         struct Sprite   *sprite     = (sprite_arr   + order);
         struct Position *position   = (position_arr + order);
@@ -162,7 +162,7 @@ void Draw_Map_HPBar(tnecs_input *input) {
     struct Position *pos_arr        = TNECS_COMPONENT_ARRAY(input, Position_ID);
     struct MapHPBar *map_hp_bar_arr = TNECS_COMPONENT_ARRAY(input, MapHPBar_ID);
 
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct MapHPBar *map_hp_bar = (map_hp_bar_arr + order);
 
         if ((!map_hp_bar->visible) && (!map_hp_bar->update))
@@ -188,7 +188,7 @@ void Draw_PopUp(tnecs_input *input) {
     SDL_assert(popup_arr    != NULL);
     SDL_assert(position_arr != NULL);
 
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct PopUp *popup = (popup_arr + order);
 
         if (!popup->visible) /* Skip drawing if not visible */
@@ -210,7 +210,7 @@ void Draw_Menu(tnecs_input *input) {
     SDL_assert(mc_arr != NULL);
 
     /* --- DRAWING MENU ENTITIES --- */
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Menu *mc = (mc_arr + order);
 
         if (!mc->visible)
@@ -237,7 +237,7 @@ void Draw_Text(tnecs_input *input) {
     SDL_assert(position_arr != NULL);
 
     /* --- DRAWING TEXT ENTITIES --- */
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Position *pos  = (position_arr    + order);
         struct Text     *text = (text_arr        + order);
 
@@ -264,7 +264,7 @@ void Draw_Text_Timer(tnecs_input *input) {
     SDL_assert(position_arr != NULL);
 
     /* --- DRAWING TEXT ENTITIES --- */
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Position *pos  = (position_arr    + order);
         struct Text     *text = (text_arr        + order);
         struct Timer    *ut   = (updatetimer_arr + order);
@@ -298,7 +298,7 @@ void Animate_Combat_onMap(tnecs_input *input) {
     struct Timer            *timer_arr      = TNECS_COMPONENT_ARRAY(input, Timer_ID);
 
     /* --- DRAWING TEXT ENTITIES --- */
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct CombatAnimation *map_anim        = &combatanim_arr[order];
         struct Timer           *combat_timer    = &timer_arr[order];
 
@@ -321,7 +321,7 @@ void Animate_Map_Animation(tnecs_input *input) {
     struct MapAnimation *mapanim_arr    = TNECS_COMPONENT_ARRAY(input, MapAnimation_ID);
 
     /* --- DRAWING TEXT ENTITIES --- */
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct MapAnimation *map_anim   = &mapanim_arr[order];
         struct Timer        *timer      = &timer_arr[order];
 
@@ -344,7 +344,7 @@ void Animate_Unit_Move_onMap(tnecs_input *input) {
     struct UnitMoveAnimation    *unitmove_arr   = TNECS_COMPONENT_ARRAY(input, UnitMoveAnimation_ID);
 
     /* --- DRAWING TEXT ENTITIES --- */
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct UnitMoveAnimation *unit_anim  = &unitmove_arr[order];
         struct Timer             *timer      = &timer_arr[order];
 
@@ -367,7 +367,7 @@ void Animate_Cutscene(tnecs_input *input) {
     struct Cutscene *cutscene_arr   = TNECS_COMPONENT_ARRAY(input, Cutscene_ID);
 
     /* --- DRAWING TEXT ENTITIES --- */
-    for (u16 order = 0; order < input->num_entities; order++) {
+    for (size_t order = 0; order < input->num_entities; order++) {
         struct Timer    *timer      = &timer_arr[order];
         struct Cutscene *cutscene   = &cutscene_arr[order];
